Built More Foundation
A lot more has been done, including a working debug console (that needs to be turned off before release, and have that get sent to a file).

There's now basic sprite code that works.

ErrorEnum.h is there to replace the error defines in previous versions of stdafx.h so that additions are easier to manage, as well as reducing compile time.

TODO: Create a scale based system for the screen based off of the screen ratio.