Added Basic Controller Support
Added basic controller support. This includes added button support and axies support. Currently movement uses the left axis.

The D-Pad isn't working, but I don't know if this is software or hardware. I'll have to try it out with a different controller.

There is a bug where after a controller is unplugged and plugged back in, the axies and keyboard stops working.