Massive Changes to Render Engine
Render engine now uses SDL2-GPU instead of SDL2, which is faster, and allows for more complex and lower level things like shaders (which is neccessary for the upcoming HUD design).

The only difference in the effects of Sprite objects is that now 1 unit is equal on both axis, so a width is now 1.333 units.

Also switched from 32 bit SDL2_gpu binary to a 64 bit SDL_gpu binary (because of a compiler accident).