Added Release Mode
Made it to where Release mode has proper setup, and to where in release mode, it will put the debug output to file instead of to a debug console.

Release mode has high optimization making it run a lot faster and have a smaller .exe size, as opposed to debug that has none (which allows for line by line debug).