Cutscenes Added
Added cutscenes! This basically just works with json files loading up images for different durations. The test cutscene is weird because it was the first thing I found on Google Images of the correct aspect ratio (just looking up screensavers).

This also made implementations of game states (which are necessary to ensure working cutscenes), which previously was just an unused Enum.

The addition of game states also added a pause button to the game (currently Esc), which doesn't have an interface yet.

Also made various small changes to performance, executable size, readability, etc.