Added Skippable Cutscenes
Added skip button to cutscenes. This should add to the replayability of the game as well as being a feature that's expected in modern gaming. It also has it to where if the button is held for longer than half a second, it will skip on frame one (good for fast skipping).

Also fixed bug where cutscene wouldn't display properly if player moved Arena camera beforehand.