A* Corrections & Major Bug Fixes
Working on getting A* to work (still doesn't work properly).

Added AI flag to stay still if on player.

Added hitbox rectangle for debug mode to be able to see hitboxes better. This will help with matching and debugging (possibly will add an option for them to display in release for game mastery).

Properly matched hitbox to player and enemies (this is an important change), with leniancy so that the hitbox is about a pixel smaller so that you are more likely for it to look like it hits and not than for it to look like it doesn't and does.