Added Dialogue to Cutscenes and Added Sprint-Dash
Added a dialogue system to the game that allows for text to be displayed while the images go by. The size ratios may need to be changed at some point for better aethstetics

Added  a sprint-dash mechanic that allows for a short speed boost and invincibility frames (made for skillful dodging of enemies).