Added Mem Leak Detection, Fixed Mem Leaks, and Added Working Enemy Damage & Death
Added Visual Studio's memory leak detection to project. Because of this, many memory leaks were fixed. There are still 3 that always happen, but it appears to be library code (not from project files), likely an error in using rapidjson. TODO: Figure out what that is and fix it.

Enemies can now take damage, and if they're killed, removed from linked list. Currently the player can hold RSHIFT, which causes enemies touching the player (bigger than the hitbox) to be damaged. This may be kept as an interesting mechanic (likely with a key change) or may be removed later.