Code Reformatting & Major Refactoring
Did a lot of code reformatting to the new code formtting standards for the project. This should add to readability.

Did major refactoring for readability, reusability, and possibly a performance boost. I still probably need to do a bit more refactoring in the near future, but the major things have been changed.