Added "The Damned" Enemy
Added basic enemy code for The Damned, the basic enemies in Hell. This includes the modifications of variable damage for enemies so that the Killer Shadows can deal more damage.

Also moved a few files to the correct location and corrected a bit of formatting in th KillerShadow.cpp file.